+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v0.1
 
+### 2021/02/21
+- use calendar component
+
 ### 2021/02/20
 - init theme config